[xsp] Output all early errors to stderr
If an error occurs during early phase of mod-mono-server startup (including exceptions while
parsing config files) make sure the error and/or exception is output to standard error. The
reason for this is that mod_mono by default redirects stdout to /dev/null (it doesn't do that if
the MonoDebug directive is set to 'true').
Also added an internal class to safely retrieve AppSettings from the config file.