[xsp] Fix for bug #575600. MonoWorkerRequest.MapPath with multiple apps should use physical server root.
When xsp is configured to serve two virtual applications and one of them attempts to map a file from
the other, the returned physical path must be composed using the XSP's physical root as the base path - not
the current application's physical path. This is consistent with the way IIS works.
To achieve this, the commit adds a new PhysicalRoot property to the ApplicationServer class. The property is
used in MonoWorkerRequest.MapPath ().