[xsp] Move a few XSP classes back to Mono.WebServer.dll
There exist 3rd party application which use xsp.exe classes for hosting. Those
classes used to be part of Mono.WebServer.dll, which was a design mistake, and were
moved out of it to xsp.exe back in March 2010. This commit puts them back in the dll
but marks them as [Obsolete]. The classes will be removed from the dll when Mono 3.0
is released.