[xsp] Improved single app mode shutdown sequence.
When XSP/mod-mono-server are operating in the single app mode, the application
runs in the same AppDomain what the Server instance. This requires special attention
to handling of the AppDomain's unload. This commit adds code which waits for the
AppDomain unload process to complete before recreating the domain. Without this code
it may happen that the main listening socket will not be destroyed in the "old" domain
before the "new" domain attempts to bind its own listening socket to the same port (which
leads to bind failure and exit from the process). The code is not very elegant but it is
portable.