[xsp] Be more careful with memory allocation in ModMonoRequest.FillBuffer and dispose the worker properly
FillBuffer will not blindly allocate the byte array with the size equal to the count value passed
to it. Instead it uses a maximum of 128k for the fill_buffer array and reads the data in chunks,
should the requested count exceed the configured maximum.
Additionally, ModMonoApplicationHost.ProcessRequest catches early errors in ModMonoWorkerRequest and
disposes of both the request and the worker by calling EndOfRequest.
ModMonoRequest and ModMonoWorker are now IDisposable.