[FastCgi] Don't delete the socket file if it doesn't match the original
A good practice for deploying a new version of a web app is to do it over
the same unix domain socket of the previous release (overwriting it). In
this case, it is not desirable that Mono.FastCgi deletes the socket file
when the previous release is killed, as it would remove the socket file
of the new release.

The best way to check if we can delete the socket file is to store the
inode of the file when the unix domain socket file is created, and check
that it is the same before deleting it, at dispose time.