Fixes wrong addition of elapsed time during lingering close.
On each iteration receiving lingering data, the absolute time
elapsed from the begining of the loop was being added, instead
of only the time spent during current iteration.

This caused the loop to finishing to soon, hence provoking a
premature close of the socket, by which the client receives
some TCP/RSTs instead of a gracefull connection shutdown.